Ask a confirmation before deleting animal.

+++ client/src/components/Animal.js
@@ -30,17 +30,18 @@ const Animal = ({
   };
   const handleDelete = e => {
     e.preventDefault();
-    DeleteAnimal(AnimalId)
-      .then(res => {
-        if (res.status === 204) {
-          window.alert("Animal Deleted!");
-        } else {
-          window.alert("Error " + res.status);
-        }
-      })
-      .catch(err => {
-        window.alert("Error " + err.response.status);
-      });
+    if (window.confirm("Are you sure you want to delete?"))
+      DeleteAnimal(AnimalId)
+        .then(res => {
+          if (res.status === 204) {
+            window.alert("Animal Deleted!");
+          } else {
+            window.alert("Error " + res.status);
+          }
+        })
+        .catch(err => {
+          window.alert("Error " + err.response.status);
+        });
   };
   return (
     <div className="Animals-Item py-3 col-4">
